refactor: relocate old Operate importer to importer-8_7
