+++ operate/importer/src/test/java/io/camunda/operate/zeebeimport/ImportListenerIT.java
@@ -25,7 +25,7 @@ import io.camunda.operate.store.elasticsearch.ElasticsearchZeebeStore;
 import io.camunda.operate.util.NoBeansIT;
 import io.camunda.operate.util.apps.nobeans.TestApplicationWithNoBeans;
 import io.camunda.operate.zeebe.ImportValueType;
-import io.camunda.operate.zeebeimport.v8_6.processors.ImportBulkProcessor;
+import io.camunda.operate.zeebeimport.v8_7.processors.processors.ImportBulkProcessor;
 import io.camunda.webapps.schema.entities.operate.ImportPositionEntity;
 import java.util.ArrayList;
 import org.elasticsearch.client.RestHighLevelClient;
