+++ operate/pom.xml
@@ -20,7 +20,7 @@
     <module>schema</module>
     <module>data-generator</module>
     <module>importer-common</module>
-    <module>importer-8_6</module>
+    <module>importer-8_7</module>
     <module>importer</module>
     <module>webapp</module>
     <module>mvc-auth-commons</module>
