+++ parent/pom.xml
@@ -598,7 +598,7 @@
 
       <dependency>
         <groupId>io.camunda</groupId>
-        <artifactId>operate-importer-8_6</artifactId>
+        <artifactId>operate-importer-8_7</artifactId>
         <version>${project.version}</version>
       </dependency>
